Add "Identity Mappings in Deep Residual Networks"
"Identity Mappings in Deep Residual Networks" is a follow up on "Deep residual learning for image recognition" in which the authors expand their ideas. They are able to effectively train a 1000 layer deep CNN on CIFAR-10 and CIFAR-100 achieving state of the art results. Due to the current lack of citations it only qualifies as a paper worth reading, but it is already mentioned in the course notes of Stanford's CS231n: https://cs231n.github.io/convolutional-networks/#case